requirements.txt
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==20.3.0
Authlib==0.15.4
autopep8==2.0.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==6.0.0
cachetools==4.2.1
celery==4.4.7
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==5.5
cron-descriptor==1.2.35
cryptography==3.4.7
defusedxml==0.7.0
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.6
Django==3.2.19
django-active-link==0.1.8
django-admin-rangefilter==0.6.4
django-admin-sortable2==1.0
django-anymail==8.1
django-appconf==1.0.4
django-author==1.2.0
django-auto-logout==0.5.0
django-celery-beat==2.0.0
django-celery-email==3.0.0
django-celery-results==1.2.1
django-constance==2.7.0
django-cors-headers==3.5.0
django-countries==7.0
django-coverage-plugin==1.8.0
django-debug-toolbar==3.1.1
django-dotenv==1.4.2
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
django-import-export==2.5.0
django-import-export-celery==1.3
django-mail-panel==3.1
django-model-utils==4.0.0
django-modeltranslation==0.16
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-ordered-model==3.4.1
django-otp==1.1.1
django-phonenumber-field==5.0.0
django-phonenumbers==1.0.1
django-picklefield==3.0.1
django-rangefilter==0.9.1
django-redis==5.2.0
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
django-simple-history==3.1.1
django-summernote==0.8.11.6
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
et-xmlfile==1.0.1
factory-boy==3.2.1
Faker==15.1.1
fastdiff==0.3.0
filelock==3.12.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.15.0
graphene-django-optimizer==0.8.0
graphene-file-upload==1.3.0
graphene-permissions==1.1.4
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.19.0
idna==2.10
imageio==2.30.0
importlib-metadata==4.12.0
iniconfig==1.1.1
jdcal==1.4.1
kombu==4.6.11
Markdown==3.4.1
MarkupPy==1.14
mongoengine==0.27.0
MyQR==2.3.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
ordered-set==4.0.2
packaging==20.9
phonenumbers==8.13.9
phonenumberslite==8.12.13
Pillow==9.5.0
platformdirs==3.2.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.38
protobuf==3.15.5
psycopg2==2.9.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.20
pyee==8.2.2
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==2.4.7
PyPDF2==3.0.1
pypng==0.20220715.0
pyppeteer==0.2.6
pytest==6.2.5
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2021.1
redis==4.3.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
sentry-sdk==0.19.2
setuptools-git==1.2
singledispatch==3.6.1
six==1.15.0
snapshottest==0.6.0
soupsieve==2.2
sqlparse==0.2.4
tablib==3.0.0
termcolor==1.1.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
virtualenv==20.21.0
wasmer==1.1.0
wasmer-compiler-cranelift==1.1.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==9.1
wrapt==1.14.1
xero-python==1.8.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
django_extensions
whitenoise
user_agents
pyfcm
geocoder